# public/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Source Library API",
    "description": "Search and cite rare historical texts (alchemy, Hermeticism, Renaissance philosophy) from Source Library. All translations include DOI-backed academic citations.",
    "version": "1.1.0",
    "contact": {
      "name": "Source Library",
      "url": "https://sourcelibrary.org",
      "email": "team@sourcelibrary.org"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://sourcelibrary.org/api",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/books/{id}": {
      "get": {
        "operationId": "getBook",
        "summary": "Get book details",
        "description": "Get detailed information about a specific book including summary, edition info, and DOI.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Book ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDetail"
                }
              }
            }
          },
          "404": {
            "description": "Book not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "display_title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "published": {
            "type": "string"
          },
          "publisher": {
            "type": "string"
          },
          "place_published": {
            "type": "string"
          },
          "pages_count": {
            "type": "integer"
          },
          "pages_translated": {
            "type": "integer"
          },
          "doi": {
            "type": "string"
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "summary": {
            "type": "string"
          },
          "editions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Edition"
            }
          }
        }
      },
      "Edition": {
        "type": "object",
        "properties": {
          "version": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ]
          },
          "doi": {
            "type": "string"
          },
          "published_at": {
            "type": "string",
            "format": "date-time"
          },
          "license": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Dataset API key (sl_data_...). Free at https://sourcelibrary.org/developers."
      }
    }
  }
}